Update plugin framework
