+++ pyload/plugin/Hook.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 
-from pyload.plugin.Addon import Addon, threaded
+from pyload.plugin.Addon import Addon, Hoster, threaded
 
 
-class Hook(Addon):
+class Hook(Addon, Hoster):
     __name    = "Hook"
     __type    = "hook"
     __version = "0.03"
