+++ pyload/plugin/hoster/TurbobitNet.py
@@ -97,10 +97,10 @@ class TurbobitNet(SimpleHoster):
 
 
     def getRtUpdate(self):
-        rtUpdate = self.getStorage("rtUpdate")
+        rtUpdate = self.retrieve("rtUpdate")
         if not rtUpdate:
-            if self.getStorage("version") != self.__version \
-               or int(self.getStorage("timestamp", 0)) + 86400000 < timestamp():
+            if self.retrieve("version") != self.__version \
+               or int(self.retrieve("timestamp", 0)) + 86400000 < timestamp():
                 # that's right, we are even using jdownloader updates
                 rtUpdate = getURL("http://update0.jdownloader.org/pluginstuff/tbupdate.js")
                 rtUpdate = self.decrypt(rtUpdate.splitlines()[1])
@@ -109,9 +109,9 @@ class TurbobitNet(SimpleHoster):
                                   r'zza=\2;for(var zzi=0;zzi<zza.length;zzi++){\1=zza[zzi];', rtUpdate)
                 rtUpdate = re.sub(r"for\((\w+)=", r"for(var \1=", rtUpdate)
 
-                self.setStorage("rtUpdate", rtUpdate)
-                self.setStorage("timestamp", timestamp())
-                self.setStorage("version", self.__version)
+                self.store("rtUpdate", rtUpdate)
+                self.store("timestamp", timestamp())
+                self.store("version", self.__version)
             else:
                 self.logError(_("Unable to download, wait for update..."))
                 self.tempOffline()
